refactor(profile): split into hero and settings components
