+++ frontend/src/app/features/profile/profile-hero/profile-hero.component.html
@@ -1,0 +1,22 @@
+<div class="hero-section">
+  <div class="hero-background">
+    <div class="hero-gradient"></div>
+    <div class="hero-pattern"></div>
+  </div>
+  <div class="hero-content">
+    <div class="profile-avatar">
+      <span>{{ getInitial() }}</span>
+    </div>
+    <div class="profile-info">
+      <h1 class="profile-name">{{ getDisplayName() }}</h1>
+      <p class="profile-email">{{ user!.email }}</p>
+      <p class="profile-join-date">
+        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+            d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z" />
+        </svg>
+        Member since {{ user!.createdAt | date:'mediumDate' }}
+      </p>
+    </div>
+  </div>
+</div>
